Extract a character column into multiple columns using regular expression groups

Given a regular expression with capturing groups, `extract()` turns each group into a new column. If the groups don't match, or the input is NA, the output will be NA.

extract(
  data,
  col,
  into,
  regex = "([[:alnum:]]+)",
  remove = TRUE,
  convert = FALSE,
  ...
)

Arguments

data
A tidySE object

col
Column name or position. This is passed to [tidyselect::vars_pull()].

This argument is passed by expression and supports [quasiquotation][rlang::quasiquotation] (you can unquote column names or column positions).

into
Names of new variables to create as character vector. Use `NA` to omit the variable in the output.

regex
a regular expression used to extract the desired values. There should be one group (defined by `()`) for each element of `into`.

remove
If `TRUE`, remove input column from output data frame.

convert
If `TRUE`, will run [type.convert()] with `as.is=TRUE` on new columns. This is useful if the component columns are integer, numeric or logical.

NB: this will cause string `"NA"`s to be converted to `NA`s.

...
Additional arguments passed on to methods.

Value

A tidySE objector a tibble depending on input

See also

Examples

tidySE::pasilla %>%
    tidy() %>%
    extract(type, into="sequencing", regex="([a-z]*)_end", convert=TRUE)

#> # A tibble: 102,193 x 5
#>    sample condition sequencing transcript  counts
#>    <chr>  <chr>     <chr>      <chr>        <int>
#>  1 untrt1 untreated single     FBgn0000003      0
#>  2 untrt1 untreated single     FBgn0000008     92
#>  3 untrt1 untreated single     FBgn0000014      5
#>  4 untrt1 untreated single     FBgn0000015      0
#>  5 untrt1 untreated single     FBgn0000017   4664
#>  6 untrt1 untreated single     FBgn0000018    583
#>  7 untrt1 untreated single     FBgn0000022      0
#>  8 untrt1 untreated single     FBgn0000024     10
#>  9 untrt1 untreated single     FBgn0000028      0
#> 10 untrt1 untreated single     FBgn0000032   1446
#> # … with 102,183 more rows